requirements.txt
ase
h5py
numpy
tensorflow
pyyaml
